Split model into meta and content files
