+++ src/fmu/dataio/models/meta/__init__.py
@@ -1,0 +1,6 @@
+from .meta import Root, dump
+
+__all__ = [
+    "dump",
+    "Root",
+]
