+++ src/fmu/dataio/models/meta/__main__.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+import json
+
+from . import dump
+
+if __name__ == "__main__":
+    print(json.dumps(dump(), indent=2))
